requirements.txt
appnope==0.1.0
boto==2.48.0
decorator==4.0.11
dj-database-url==0.4.2
Django==2.1.1
django-filter==1.0.4
django-letsencrypt==3.0.0
django-markdownx==2.0.22
django-prerenderio==17.6.1
django-seo-js==0.3.1
django-storages==1.6.5
djangorestframework==3.8.2
gunicorn==19.7.1
ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
Markdown==2.6.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.3
requests==2.2.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
whitenoise==3.3.0
